+++ src/main/java/com/troblecodings/signals/guis/GuiSignalBox.java
@@ -13,6 +13,7 @@ import java.util.stream.Collectors;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.troblecodings.core.I18Wrapper;
+import com.troblecodings.core.TCBoolean;
 import com.troblecodings.core.WriteBuffer;
 import com.troblecodings.guilib.ecs.ContainerBase;
 import com.troblecodings.guilib.ecs.DrawUtil.DisableIntegerable;
@@ -76,16 +77,17 @@ public class GuiSignalBox extends GuiBase {
     public static final int GRID_COLOR = 0xFF5B5B5B;
     public static final int EDIT_COLOR = 0x5000A2FF;
     public static final int OUTPUT_COLOR = 0xffff00;
-    public static final int TRAIN_NUMBER_BACKGROUND_COLOR = ConfigHandler.signalboxTrainnumberBackgroundColor;
+    public static final int TRAIN_NUMBER_BACKGROUND_COLOR =
+            ConfigHandler.signalboxTrainnumberBackgroundColor;
 
-    public static final ResourceLocation REDSTONE_OFF = new ResourceLocation(OpenSignalsMain.MODID,
-            "gui/textures/redstone_off.png");
-    public static final ResourceLocation REDSTONE_OFF_BLOCKED = new ResourceLocation(
-            OpenSignalsMain.MODID, "gui/textures/redstone_off_blocked.png");
-    public static final ResourceLocation REDSTONE_ON = new ResourceLocation(OpenSignalsMain.MODID,
-            "gui/textures/redstone_on.png");
-    public static final ResourceLocation REDSTONE_ON_BLOCKED = new ResourceLocation(
-            OpenSignalsMain.MODID, "gui/textures/redstone_on_blocked.png");
+    public static final ResourceLocation REDSTONE_OFF =
+            new ResourceLocation(OpenSignalsMain.MODID, "gui/textures/redstone_off.png");
+    public static final ResourceLocation REDSTONE_OFF_BLOCKED =
+            new ResourceLocation(OpenSignalsMain.MODID, "gui/textures/redstone_off_blocked.png");
+    public static final ResourceLocation REDSTONE_ON =
+            new ResourceLocation(OpenSignalsMain.MODID, "gui/textures/redstone_on.png");
+    public static final ResourceLocation REDSTONE_ON_BLOCKED =
+            new ResourceLocation(OpenSignalsMain.MODID, "gui/textures/redstone_on_blocked.png");
 
     private static final float[] ALL_LINES = getLines();
     protected static final int TILE_WIDTH = 10;
@@ -126,6 +128,9 @@ public class GuiSignalBox extends GuiBase {
     protected final Map<BlockPos, SubsidiaryHolder> enabledSubsidiaries = new HashMap<>();
     private final Map<Point, UIColor> colors = new HashMap<>();
 
+    private int index = 0;
+    private List<Point> pointList = new ArrayList<>();
+
     public GuiSignalBox(final GuiInfo info) {
         super(info);
         this.container = (ContainerSignalBox) info.base;
@@ -134,6 +139,7 @@ public class GuiSignalBox extends GuiBase {
         container.setSignalUpdater(this::updateSignals);
         container.setConuterUpdater(this::updateCounter);
         container.setTrainNumberUpdater(this::updateTrainNumber);
+        container.setDebugPointUpdater(this::setDebugPoint);
         this.info = info;
     }
 
@@ -233,8 +239,8 @@ public class GuiSignalBox extends GuiBase {
 
     public static String getSignalInfo(final BlockPos signalPos, final LinkType type) {
         final Minecraft mc = Minecraft.getMinecraft();
-        final String customName = ClientNameHandler
-                .getClientName(new StateInfo(mc.world, signalPos));
+        final String customName =
+                ClientNameHandler.getClientName(new StateInfo(mc.world, signalPos));
         return String.format("%s (x=%d, y=%d. z=%d)", customName == null
                 ? (type.equals(LinkType.SIGNAL) ? "" : I18Wrapper.format("type." + type.name()))
                 : customName, signalPos.getX(), signalPos.getY(), signalPos.getZ());
@@ -248,9 +254,9 @@ public class GuiSignalBox extends GuiBase {
         helpPage.helpUsageMode(null);
         this.resetTileSelection();
 
-        final MainSignalIdentifier identifier = new MainSignalIdentifier(
-                new ModeIdentifier(holder.point, holder.modeSet), pos,
-                SignalState.combine(entry.enumValue.getSubsidiaryShowType()));
+        final MainSignalIdentifier identifier =
+                new MainSignalIdentifier(new ModeIdentifier(holder.point, holder.modeSet), pos,
+                        SignalState.combine(entry.enumValue.getSubsidiaryShowType()));
         final List<MainSignalIdentifier> greenSignals = container.greenSignals
                 .computeIfAbsent(identifier.getPoint(), _u -> new ArrayList<>());
         greenSignals.remove(identifier);
@@ -302,7 +308,10 @@ public class GuiSignalBox extends GuiBase {
                 }
             }
         }));
-        tile.add(new UIClickable(e -> openNodeShortcuts(currentTile.getNode(), e), 1));
+        tile.add(new UIClickable(e -> {
+            openNodeShortcuts(currentTile.getNode(), e);
+            getDebugPoint();
+        }, 1));
     }
 
     private void checkForMultiplePathTypes(final SignalBoxNode start, final SignalBoxNode end) {
@@ -379,8 +388,8 @@ public class GuiSignalBox extends GuiBase {
         nameEntity.setHeight(20);
         nameEntity.add(new UIBox(UIBox.HBOX, 5));
 
-        final UIEntity labelEntity = GuiElements.createLabel(I18Wrapper.format("info.node.text"),
-                1.25f);
+        final UIEntity labelEntity =
+                GuiElements.createLabel(I18Wrapper.format("info.node.text"), 1.25f);
         labelEntity.setInheritWidth(false);
         labelEntity.setWidth(100);
         nameEntity.add(labelEntity);
@@ -562,8 +571,8 @@ public class GuiSignalBox extends GuiBase {
                 bottomEntity.add(menu);
                 bottomEntity.getParent().update();
             });
-            final UIEntity buttonNo = GuiElements.createButton(I18Wrapper.format("btn.no"),
-                    e -> pop());
+            final UIEntity buttonNo =
+                    GuiElements.createButton(I18Wrapper.format("btn.no"), e -> pop());
             buttons.setInherits(true);
             final UIBox vbox = new UIBox(UIBox.HBOX, 1);
             buttons.add(vbox);
@@ -672,8 +681,8 @@ public class GuiSignalBox extends GuiBase {
                 this::initializePageSettings));
         header.add(
                 GuiElements.createButton(I18Wrapper.format("btn.edit"), this::initializeFieldEdit));
-        mainButton = GuiElements.createButton(I18Wrapper.format("btn.main"),
-                this::initializeFieldUsage);
+        mainButton =
+                GuiElements.createButton(I18Wrapper.format("btn.main"), this::initializeFieldUsage);
         header.add(mainButton);
         resetSelection(mainButton);
 
@@ -761,6 +770,20 @@ public class GuiSignalBox extends GuiBase {
         final WriteBuffer buffer = new WriteBuffer();
         buffer.putEnumValue(SignalBoxNetwork.SEND_ZS2_ENTRY);
         buffer.putByte(value);
+        node.getPoint().writeNetwork(buffer);
+        buffer.putByte((byte) mode.ordinal());
+        buffer.putByte((byte) rotation.ordinal());
+        buffer.putByte((byte) entry.getID());
+        OpenSignalsMain.network.sendTo(info.player, buffer);
+    }
+
+    protected void sendZS6Entry(final boolean value, final SignalBoxNode node,
+            final EnumGuiMode mode, final Rotation rotation, final PathEntryType<TCBoolean> entry) {
+        if (!allPacketsRecived)
+            return;
+        final WriteBuffer buffer = new WriteBuffer();
+        buffer.putEnumValue(SignalBoxNetwork.SEND_ZS6_ENTRY);
+        buffer.putBoolean(value);
         node.getPoint().writeNetwork(buffer);
         buffer.putByte((byte) mode.ordinal());
         buffer.putByte((byte) rotation.ordinal());
@@ -999,7 +1022,8 @@ public class GuiSignalBox extends GuiBase {
         nodes.forEach(node -> {
             final UISignalBoxTile tile = allTiles.get(node.getPoint());
             node.forEach(mode -> {
-                if (!(mode.mode == EnumGuiMode.STRAIGHT || mode.mode == EnumGuiMode.CORNER))
+                if (!(mode.mode == EnumGuiMode.STRAIGHT || mode.mode == EnumGuiMode.CORNER
+                        || mode.mode == EnumGuiMode.CROSSING))
                     return;
                 if (node.containsManuellOutput(mode)) {
                     tile.setColor(mode, OUTPUT_COLOR);
@@ -1021,6 +1045,7 @@ public class GuiSignalBox extends GuiBase {
                 switch (guiMode) {
                     case STRAIGHT:
                     case CORNER:
+                    case CROSSING:
                         tile.setColor(mode, SignalBoxUtil.FREE_COLOR);
                         entry.getEntry(PathEntryType.PATHUSAGE).ifPresent(
                                 _u -> entry.setEntry(PathEntryType.PATHUSAGE, EnumPathUsage.FREE));
@@ -1051,5 +1076,31 @@ public class GuiSignalBox extends GuiBase {
                     .ifPresent(poe -> uiTile.setColor(modeSet, poe.getEntry(PathEntryType.PATHUSAGE)
                             .orElseGet(() -> EnumPathUsage.FREE).getColor()));
         }
+    }
+
+    private void setDebugPoint(final List<Point> points) {
+        pointList = points;
+    }
+
+    private void getDebugPoint() {
+        if (index >= pointList.size()) {
+            pointList.clear();
+            index = 0;
+            return;
+        }
+        Point point = pointList.get(index);
+        System.out.println(point.getX() + ", " + point.getY());
+        SignalBoxNode node = container.grid.getNode(point);
+        if (node == null)
+            return;
+        node.forEach(mode -> {
+            if (!(mode.mode == EnumGuiMode.STRAIGHT || mode.mode == EnumGuiMode.CORNER
+                    || mode.mode == EnumGuiMode.CROSSING))
+                return;
+            final UISignalBoxTile tile = allTiles.get(node.getPoint());
+            UIColor color = new UIColor(0x4016fffe);
+            tile.getParent().add(color);
+        });
+        index++;
     }
 }