+++ src/main/java/com/troblecodings/signals/tileentitys/SignalSpecialRenderer.java
@@ -1,21 +1,27 @@
 package com.troblecodings.signals.tileentitys;
 
+import com.troblecodings.signals.core.RenderAnimationInfo;
 import com.troblecodings.signals.core.RenderOverlayInfo;
 
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 
 public class SignalSpecialRenderer extends TileEntitySpecialRenderer<SignalTileEntity> {
-
+    
     @Override
-    public void render(final SignalTileEntity te, final double x, final double y, final double z,
+    public void render(final SignalTileEntity tile, final double x, final double y, final double z,
             final float partialTicks, final int destroyStage, final float alpha) {
-        if (!te.hasCustomName())
-            return;
-        te.renderOverlay(new RenderOverlayInfo(x, y, z, getFontRenderer()));
+        if (tile.hasCustomName()) {
+            tile.renderOverlay(new RenderOverlayInfo(x, y, z, getFontRenderer()));
+        }
+        if (tile.hasAnimation()) {
+            tile.getAnimationHandler().render(new RenderAnimationInfo(x, y, z,
+                    Minecraft.getMinecraft().getRenderPartialTicks()).with(tile));
+        }
     }
 
     @Override
     public boolean isGlobalRenderer(final SignalTileEntity te) {
-        return te.hasCustomName();
+        return te.hasCustomName() || te.hasAnimation();
     }
 }